Split input field into seperate widgets
