+++ lib/overlay/overlay_ui.dart
@@ -19,7 +19,7 @@ import 'package:rxdart/subjects.dart';
 import 'package:widget_and_text_animator/widget_and_text_animator.dart';
 import 'package:window_manager/window_manager.dart';
 
-import '../widgets/input_field.dart';
+import '../widgets/input_field/input_field.dart';
 import '../widgets/markdown_builders/code_wrapper.dart';
 import '../widgets/message_list_tile.dart';
