+++ lib/overlay/search_overlay_ui.dart
@@ -16,7 +16,9 @@ import 'package:fluent_gpt/theme.dart';
 import 'package:fluent_gpt/tray.dart';
 import 'package:fluent_gpt/utils.dart';
 import 'package:fluent_gpt/widgets/drop_region.dart';
-import 'package:fluent_gpt/widgets/input_field.dart';
+import 'package:fluent_gpt/widgets/input_field/additional_btns_input_field.dart';
+import 'package:fluent_gpt/widgets/input_field/aliases_overlay.dart';
+import 'package:fluent_gpt/widgets/input_field/input_field.dart';
 import 'package:fluent_gpt/widgets/markdown_builders/code_wrapper.dart';
 import 'package:fluent_gpt/widgets/message_list_tile.dart';
 import 'package:fluentui_system_icons/fluentui_system_icons.dart' as ic;
